+++ tests/test_plugin.py
@@ -30,6 +30,7 @@ import xarray as xr
 
 import pandora
 from pandora import stereo, optimization
+from pandora.state_machine import PandoraMachine
 
 
 class TestPlugin(unittest.TestCase):
@@ -41,16 +42,16 @@ class TestPlugin(unittest.TestCase):
         Method called to prepare the test fixture
 
         """
-        self.ref = pandora.read_img('tests/ref.png', no_data=np.nan, cfg={'nodata1': 'np.nan', 'nodata2': 'np.nan',
+        self.left = pandora.read_img('tests/left.png', no_data=np.nan, cfg={'nodata1': 'np.nan', 'nodata2': 'np.nan',
                                                                           'valid_pixels': 0, 'no_data': 1}, mask=None)
-        self.sec = pandora.read_img('tests/sec.png', no_data=np.nan, cfg={'nodata1': 'np.nan', 'nodata2': 'np.nan',
+        self.right = pandora.read_img('tests/right.png', no_data=np.nan, cfg={'nodata1': 'np.nan', 'nodata2': 'np.nan',
                                                                           'valid_pixels': 0, 'no_data': 1}, mask=None)
-        self.disp_ref = rasterio.open('tests/disp_ref.tif').read(1)
-        self.disp_sec = rasterio.open('tests/disp_sec.tif').read(1)
+        self.disp_left = rasterio.open('tests/disp_left.tif').read(1)
+        self.disp_right = rasterio.open('tests/disp_right.tif').read(1)
         self.occlusion = rasterio.open('tests/occl.png').read(1)
 
-        self.disp_ref_zncc = rasterio.open('tests/disp_ref_zncc.tif').read(1)
-        self.disp_sec_zncc = rasterio.open('tests/disp_sec_zncc.tif').read(1)
+        self.disp_left_zncc = rasterio.open('tests/disp_left_zncc.tif').read(1)
+        self.disp_right_zncc = rasterio.open('tests/disp_right_zncc.tif').read(1)
 
     def error(self, data, gt, threshold, unknown_disparity=0):
         """
@@ -102,50 +103,42 @@ class TestPlugin(unittest.TestCase):
         """
         user_cfg = pandora.read_config_file('conf/sgm.json')
 
-        # Load validation plugin
-        validation_ = pandora.validation.AbstractValidation(**user_cfg["validation"])
-
-        # Build the default configuration
-        cfg = pandora.JSON_checker.default_short_configuration
-        cfg_validation = {'validation': validation_.cfg}
-        cfg = pandora.JSON_checker.update_conf(cfg, cfg_validation)
-
-        # Update the configuration with default values
-        cfg = pandora.JSON_checker.update_conf(cfg, user_cfg)
-
         # Import pandora plugins
         pandora.import_plugin()
 
+        # Instantiate machine
+        pandora_machine = PandoraMachine()
+
         # Run the pandora pipeline
-        ref, sec = pandora.run(self.ref, self.sec, -60, 0, cfg)
+        left, right = pandora.run(pandora_machine, self.left, self.right, -60, 0, user_cfg)
 
-        # Compares the calculated reference disparity map with the ground truth
+        # Compares the calculated left disparity map with the ground truth
         # If the percentage of pixel errors is > 0.20, raise an error
-        if self.error(ref['disparity_map'].data, self.disp_ref, 1) > 0.20:
+        if self.error(left['disparity_map'].data, self.disp_left, 1) > 0.20:
             raise AssertionError
 
-        # Compares the calculated reference disparity map with the ground truth
+        # Compares the calculated left disparity map with the ground truth
         # If the percentage of pixel errors ( error if ground truth - calculate > 2) is > 0.15, raise an error
-        if self.error(ref['disparity_map'].data, self.disp_ref, 2) > 0.15:
+        if self.error(left['disparity_map'].data, self.disp_left, 2) > 0.15:
             raise AssertionError
 
-        # Check the reference validity mask cross checking ( bit 8 and 9 )
+        # Check the left validity mask cross checking ( bit 8 and 9 )
         # Compares the calculated validity mask with the ground truth ( occlusion mask )
-        occlusion = np.ones((ref['validity_mask'].shape[0], ref['validity_mask'].shape[1]))
-        occlusion[ref['validity_mask'].data >= 512] = 0
+        occlusion = np.ones((left['validity_mask'].shape[0], left['validity_mask'].shape[1]))
+        occlusion[left['validity_mask'].data >= 512] = 0
 
         # If the percentage of errors is > 0.15, raise an error
         if self.error_mask(occlusion, self.occlusion) > 0.15:
             raise AssertionError
 
-        # Compares the calculated secondary disparity map with the ground truth
+        # Compares the calculated right disparity map with the ground truth
         # If the percentage of pixel errors is > 0.20, raise an error
-        if self.error(-1 * sec['disparity_map'].data, self.disp_sec, 1) > 0.20:
+        if self.error(-1 * right['disparity_map'].data, self.disp_right, 1) > 0.20:
             raise AssertionError
 
-        # Compares the calculated secondary disparity map with the ground truth
+        # Compares the calculated right disparity map with the ground truth
         # If the percentage of pixel errors ( error if ground truth - calculate > 2) is > 0.15, raise an error
-        if self.error(-1 * sec['disparity_map'].data, self.disp_sec, 2) > 0.15:
+        if self.error(-1 * right['disparity_map'].data, self.disp_right, 2) > 0.15:
             raise AssertionError
 
     def test_libsgm_negative_disparities(self):
@@ -155,50 +148,42 @@ class TestPlugin(unittest.TestCase):
         """
         user_cfg = pandora.read_config_file('conf/sgm.json')
 
-        # Load validation plugin
-        validation_ = pandora.validation.AbstractValidation(**user_cfg["validation"])
-
-        # Build the default configuration
-        cfg = pandora.JSON_checker.default_short_configuration
-        cfg_validation = {'validation': validation_.cfg}
-        cfg = pandora.JSON_checker.update_conf(cfg, cfg_validation)
-
-        # Update the configuration with default values
-        cfg = pandora.JSON_checker.update_conf(cfg, user_cfg)
-
         # Import pandora plugins
         pandora.import_plugin()
 
+        # Instantiate machine
+        pandora_machine = PandoraMachine()
+
         # Run the pandora pipeline
-        ref, sec = pandora.run(self.ref, self.sec, -60, -1, cfg)
+        left, right = pandora.run(pandora_machine, self.left, self.right, -60, -1, user_cfg)
 
-        # Compares the calculated reference disparity map with the ground truth
+        # Compares the calculated left disparity map with the ground truth
         # If the percentage of pixel errors is > 0.20, raise an error
-        if self.error(ref['disparity_map'].data, self.disp_ref, 1) > 0.20:
+        if self.error(left['disparity_map'].data, self.disp_left, 1) > 0.20:
             raise AssertionError
 
-        # Compares the calculated reference disparity map with the ground truth
+        # Compares the calculated left disparity map with the ground truth
         # If the percentage of pixel errors ( error if ground truth - calculate > 2) is > 0.15, raise an error
-        if self.error(ref['disparity_map'].data, self.disp_ref, 2) > 0.15:
+        if self.error(left['disparity_map'].data, self.disp_left, 2) > 0.15:
             raise AssertionError
 
-        # Check the reference validity mask cross checking ( bit 8 and 9 )
+        # Check the left validity mask cross checking ( bit 8 and 9 )
         # Compares the calculated validity mask with the ground truth ( occlusion mask )
-        occlusion = np.ones((ref['validity_mask'].shape[0], ref['validity_mask'].shape[1]))
-        occlusion[ref['validity_mask'].data >= 512] = 0
+        occlusion = np.ones((left['validity_mask'].shape[0], left['validity_mask'].shape[1]))
+        occlusion[left['validity_mask'].data >= 512] = 0
 
         # If the percentage of errors is > 0.15, raise an error
         if self.error_mask(occlusion, self.occlusion) > 0.15:
             raise AssertionError
 
-        # Compares the calculated secondary disparity map with the ground truth
+        # Compares the calculated right disparity map with the ground truth
         # If the percentage of pixel errors is > 0.20, raise an error
-        if self.error(-1 * sec['disparity_map'].data, self.disp_sec, 1) > 0.20:
+        if self.error(-1 * right['disparity_map'].data, self.disp_right, 1) > 0.20:
             raise AssertionError
 
-        # Compares the calculated secondary disparity map with the ground truth
+        # Compares the calculated right disparity map with the ground truth
         # If the percentage of pixel errors ( error if ground truth - calculate > 2) is > 0.15, raise an error
-        if self.error(-1 * sec['disparity_map'].data, self.disp_sec, 2) > 0.15:
+        if self.error(-1 * right['disparity_map'].data, self.disp_right, 2) > 0.15:
             raise AssertionError
 
     def test_libsgm_positive_disparities(self):
@@ -208,40 +193,32 @@ class TestPlugin(unittest.TestCase):
         """
         user_cfg = pandora.read_config_file('conf/sgm.json')
 
-        # Load validation plugin
-        validation_ = pandora.validation.AbstractValidation(**user_cfg["validation"])
-
-        # Build the default configuration
-        cfg = pandora.JSON_checker.default_short_configuration
-        cfg_validation = {'validation': validation_.cfg}
-        cfg = pandora.JSON_checker.update_conf(cfg, cfg_validation)
-
-        # Update the configuration with default values
-        cfg = pandora.JSON_checker.update_conf(cfg, user_cfg)
-
         # Import pandora plugins
         pandora.import_plugin()
 
-        sec, ref = pandora.run(self.sec, self.ref, 1, 60, cfg)
+        # Instantiate machine
+        pandora_machine = PandoraMachine()
 
-        # Compares the calculated reference disparity map with the ground truth
+        right, left = pandora.run(pandora_machine, self.right, self.left, 1, 60, user_cfg)
+
+        # Compares the calculated left disparity map with the ground truth
         # If the percentage of pixel errors is > 0.20, raise an error
-        if self.error(ref['disparity_map'].data, self.disp_ref, 1) > 0.20:
+        if self.error(left['disparity_map'].data, self.disp_left, 1) > 0.20:
             raise AssertionError
 
-        # Compares the calculated reference disparity map with the ground truth
+        # Compares the calculated left disparity map with the ground truth
         # If the percentage of pixel errors ( error if ground truth - calculate > 2) is > 0.15, raise an error
-        if self.error(ref['disparity_map'].data, self.disp_ref, 2) > 0.15:
+        if self.error(left['disparity_map'].data, self.disp_left, 2) > 0.15:
             raise AssertionError
 
-        # Compares the calculated secondary disparity map with the ground truth
+        # Compares the calculated right disparity map with the ground truth
         # If the percentage of pixel errors is > 0.20, raise an error
-        if self.error(-1 * sec['disparity_map'].data, self.disp_sec, 1) > 0.20:
+        if self.error(-1 * right['disparity_map'].data, self.disp_right, 1) > 0.20:
             raise AssertionError
 
-        # Compares the calculated secondary disparity map with the ground truth
+        # Compares the calculated right disparity map with the ground truth
         # If the percentage of pixel errors ( error if ground truth - calculate > 2) is > 0.15, raise an error
-        if self.error(-1 * sec['disparity_map'].data, self.disp_sec, 2) > 0.15:
+        if self.error(-1 * right['disparity_map'].data, self.disp_right, 2) > 0.15:
             raise AssertionError
 
     def test_libsgm_zncc(self):
@@ -251,30 +228,25 @@ class TestPlugin(unittest.TestCase):
 
         # Prepare the configuration
         user_cfg = pandora.read_config_file('conf/sgm_zncc.json')
-        cfg = pandora.JSON_checker.update_conf(pandora.JSON_checker.default_short_configuration, user_cfg)
-        cfg['input']['img_ref'] = "tests/ref.png"
-        cfg['input']['img_sec'] = "tests/sec.png"
-        cfg['input']['disp_min'] = -60
-        cfg['input']['disp_max'] = 0
 
         # Import pandora plugins
         pandora.import_plugin()
 
-        # Check the configuration
-        cfg = pandora.check_conf(cfg)
+        # Instantiate machine
+        pandora_machine = PandoraMachine()
 
         # Run the pandora pipeline
-        ref, sec = pandora.run(self.ref, self.sec, -60, 0, cfg)
+        left, right = pandora.run(pandora_machine, self.left, self.right, -60, 0, user_cfg)
 
-        # Compares the calculated reference disparity map with the ground truth
+        # Compares the calculated left disparity map with the ground truth
         # If the disparity maps are not equal, raise an error
 
-        if self.strict_error(ref['disparity_map'].data, self.disp_ref_zncc) > 0:
+        if self.strict_error(left['disparity_map'].data, self.disp_left_zncc) > 0:
             raise AssertionError
 
-        # Compares the calculated secondary disparity map with the ground truth
+        # Compares the calculated right disparity map with the ground truth
         # If the disparity maps are not equal, raise an error
-        if self.strict_error(sec['disparity_map'].data, self.disp_sec_zncc) > 0:
+        if self.strict_error(right['disparity_map'].data, self.disp_right_zncc) > 0:
             raise AssertionError
 
     def test_number_of_disp(self):
@@ -284,13 +256,12 @@ class TestPlugin(unittest.TestCase):
 
         # Prepare the configuration
         user_cfg = pandora.read_config_file('conf/sgm.json')
-        cfg = pandora.JSON_checker.update_conf(pandora.JSON_checker.default_short_configuration, user_cfg)
-        cfg['stereo']['window_size'] = 3
-        cfg['optimization']['min_cost_paths'] = True
+        user_cfg['pipeline']['stereo']['window_size'] = 3
+        user_cfg['pipeline']['optimization']['min_cost_paths'] = True
 
         # Load plugins
-        stereo_ = stereo.AbstractStereo(**cfg['stereo'])
-        optimization_ = optimization.AbstractOptimization(**cfg['optimization'])
+        stereo_ = stereo.AbstractStereo(**user_cfg['pipeline']['stereo'])
+        optimization_ = optimization.AbstractOptimization(**user_cfg['pipeline']['optimization'])
 
         # Import pandora plugins
         pandora.import_plugin()
@@ -299,19 +270,20 @@ class TestPlugin(unittest.TestCase):
                          [1, 1, 1, 1, 2],
                          [1, 1, 1, 4, 3],
                          [1, 1, 1, 1, 1]), dtype=np.float32)
-        ref = xr.Dataset({'im': (['row', 'col'], data)},
-                              coords={'row': np.arange(data.shape[0]), 'col': np.arange(data.shape[1])})
+        left = xr.Dataset({'im': (['row', 'col'], data)},
+                         coords={'row': np.arange(data.shape[0]), 'col': np.arange(data.shape[1])},
+                         attrs={'no_data_img': 0, 'valid_pixels': 0, 'no_data_mask': 1})
 
         data = np.array(([1, 1, 1, 2, 2],
                          [1, 1, 1, 4, 2],
                          [1, 1, 1, 4, 4],
                          [1, 1, 1, 1, 1]), dtype=np.float32)
-        sec = xr.Dataset({'im': (['row', 'col'], data)},
-                              coords={'row': np.arange(data.shape[0]), 'col': np.arange(data.shape[1])})
+        right = xr.Dataset({'im': (['row', 'col'], data)},
+                         coords={'row': np.arange(data.shape[0]), 'col': np.arange(data.shape[1])},
+                         attrs={'no_data_img': 0, 'valid_pixels': 0, 'no_data_mask': 1})
 
         # Computes the cost volume dataset
-        cv = stereo_.compute_cost_volume(img_ref=ref, img_sec=sec, disp_min=-2, disp_max=2,
-                                                    **{'valid_pixels': 0, 'no_data': 1})
+        cv = stereo_.compute_cost_volume(img_left=left, img_right=right, disp_min=-2, disp_max=2)
 
         # Disparities which give a minimum local cost, in indices
         disp_path = np.array([[[1, 3, 2, 3, 1, 3, 3, 2],
@@ -330,7 +302,6 @@ class TestPlugin(unittest.TestCase):
 
         # Check if the calculated confidence_measure is equal to the ground truth (same shape and all elements equals)
         np.testing.assert_array_equal(cv_updated['confidence_measure'].data[:, :, -1], gt_disp)
-
 
 
 if __name__ == '__main__':