+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+- Update depency to Pandora 0.3.0 version. 
+
 ### Fixed
 
 -  min_cost_paths calculation: fix creation of temporary disparity map to avoid dangerous DataArray manipulation. [#43]